fixed #553
