--- video_creation/data/.gitignore
@@ -1,2 +1,0 @@
-videos.json
- #todo add videos on github
